requirements.txt
--extra-index-url https://pkgs.dev.azure.com/Christian-Dior/e93e16e3-116f-4b44-a461-bf40ec428317/_packaging/CDC.pypi/pypi/simple

aiohttp==3.8.5 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
appnope==0.1.3 ; python_version >= "3.9" and python_version < "4.0" and (platform_system == "Darwin" or sys_platform == "darwin")
asttokens==2.2.1 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
backcall==0.2.0 ; python_version >= "3.9" and python_version < "4.0"
cachetools==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.9" and python_version < "4.0"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0" and implementation_name == "pypy"
charset-normalizer==3.2.0 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
comm==0.1.3 ; python_version >= "3.9" and python_version < "4.0"
contourpy==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
cycler==0.11.0 ; python_version >= "3.9" and python_version < "4.0"
db-dtypes==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
debugpy==1.6.7 ; python_version >= "3.9" and python_version < "4.0"
decorator==5.1.1 ; python_version >= "3.9" and python_version < "4.0"
executing==1.2.0 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.12.2 ; python_version >= "3.9" and python_version < "4.0"
fonttools==4.41.1 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.4.0 ; python_version >= "3.9" and python_version < "4.0"
fsspec==2023.6.0 ; python_version >= "3.9" and python_version < "4.0"
gcsfs==2023.6.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.11.1 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.11.1 ; python_version >= "3.9" and python_version < "4.0"
google-auth-oauthlib==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.22.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-bigquery-storage==2.22.0 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-bigquery==3.11.4 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.10.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.59.1 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.56.2 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.56.2 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
importlib-metadata==6.8.0 ; python_version >= "3.9" and python_version < "3.10"
importlib-resources==6.0.0 ; python_version >= "3.9" and python_version < "3.10"
ipykernel==6.25.0 ; python_version >= "3.9" and python_version < "4.0"
ipython==8.14.0 ; python_version >= "3.9" and python_version < "4.0"
jedi==0.18.2 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
jupyter-client==8.3.0 ; python_version >= "3.9" and python_version < "4.0"
jupyter-core==5.3.1 ; python_version >= "3.9" and python_version < "4.0"
kiwisolver==1.4.4 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.3 ; python_version >= "3.9" and python_version < "4.0"
matplotlib-inline==0.1.6 ; python_version >= "3.9" and python_version < "4.0"
matplotlib==3.7.2 ; python_version >= "3.9" and python_version < "4.0"
mpmath==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
nest-asyncio==1.5.6 ; python_version >= "3.9" and python_version < "4.0"
networkx==3.1 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.25.1 ; python_version >= "3.9" and python_version < "4.0"
oauthlib==3.2.2 ; python_version >= "3.9" and python_version < "4.0"
opencv-python==4.8.0.74 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.9" and python_version < "4.0"
pandas==2.0.3 ; python_version >= "3.9" and python_version < "4.0"
parso==0.8.3 ; python_version >= "3.9" and python_version < "4.0"
pexpect==4.8.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform != "win32"
pickleshare==0.7.5 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.0.0 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==3.9.1 ; python_version >= "3.9" and python_version < "4.0"
prompt-toolkit==3.0.39 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.23.4 ; python_version >= "3.9" and python_version < "4.0"
psutil==5.9.5 ; python_version >= "3.9" and python_version < "4.0"
ptyprocess==0.7.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform != "win32"
pure-eval==0.2.2 ; python_version >= "3.9" and python_version < "4.0"
py-cpuinfo==9.0.0 ; python_version >= "3.9" and python_version < "4.0"
pyarrow==12.0.1 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0" and implementation_name == "pypy"
pygments==2.15.1 ; python_version >= "3.9" and python_version < "4.0"
pyhocon==0.3.60 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
pywin32==306 ; sys_platform == "win32" and platform_python_implementation != "PyPy" and python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.9" and python_version < "4.0"
pyzmq==25.1.0 ; python_version >= "3.9" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
scipy==1.9.3 ; python_version >= "3.9" and python_version < "4.0"
seaborn==0.12.2 ; python_version >= "3.9" and python_version < "4.0"
setuptools==68.0.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
stack-data==0.6.2 ; python_version >= "3.9" and python_version < "4.0"
sympy==1.12 ; python_version >= "3.9" and python_version < "4.0"
tooling-aif-bigquery==0.2.1 ; python_version >= "3.9" and python_version < "4.0"
torch==2.0.1 ; python_version >= "3.9" and python_version < "4.0"
torchvision==0.15.2 ; python_version >= "3.9" and python_version < "4.0"
tornado==6.3.2 ; python_version >= "3.9" and python_version < "4.0"
tqdm==4.65.0 ; python_version >= "3.9" and python_version < "4.0"
traitlets==5.9.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.7.1 ; python_version >= "3.9" and python_version < "4.0"
tzdata==2023.3 ; python_version >= "3.9" and python_version < "4.0"
ultralytics==8.0.142 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.16 ; python_version >= "3.9" and python_version < "4.0"
wcwidth==0.2.6 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.9.2 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.16.2 ; python_version >= "3.9" and python_version < "3.10"
